+++ hypervisor/include/arch/x86/guest/guest.h
@@ -96,7 +96,7 @@ uint64_t vcpumask2pcpumask(struct vm *vm, uint64_t vdmask);
 int gva2gpa(struct vcpu *vcpu, uint64_t gva, uint64_t *gpa, uint32_t *err_code);
 
 struct vcpu *get_primary_vcpu(struct vm *vm);
-struct vcpu *vcpu_from_vid(struct vm *vm, int vcpu_id);
+struct vcpu *vcpu_from_vid(struct vm *vm, uint16_t vcpu_id);
 struct vcpu *vcpu_from_pid(struct vm *vm, uint16_t pcpu_id);
 
 enum vm_paging_mode get_vcpu_paging_mode(struct vcpu *vcpu);
